requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.6.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
tqdm>=4.62.0

# Deep learning
torch>=1.10.0
accelerate>=0.16.0

# NLP & transformers
transformers>=4.26.0
datasets>=2.10.0
tokenizers>=0.13.0
sentencepiece>=0.1.97
evaluate>=0.4.0
nltk>=3.7

# PEFT (Parameter-Efficient Fine-Tuning)
peft>=0.3.0

# Hyperparameter optimization
# Ray Tune is optional but recommended for advanced HPO
ray[tune]>=2.3.0
hyperopt>=0.2.7

# Utilities
pyyaml>=6.0
rich>=12.0.0
typer>=0.7.0
colorama>=0.4.4
jsonschema>=4.17.0
tabulate>=0.9.0

# Logging and experiment tracking
wandb>=0.15.0
tensorboard>=2.10.0

# Optional: HuggingFace Hub enhancements
# For better performance when downloading models and datasets
huggingface_hub[hf_xet]>=0.15.0
# Alternatively: hf_xet>=0.1.0

# Optional: model compression and optimization
bitsandbytes>=0.38.0
# optimum>=1.7.0 